sem_2/lab_15/requirements.txt
selenium
psycopg2-binary
pypika
webdriver-manager
